test: renamed dir
