+++ test/functional/fixtures/api/es-next/custom-client-scripts/common/index.js
