+++ test/functional/fixtures/api/es-next/custom-client-scripts/testcafe-fixtures/runner.js
@@ -1,4 +1,4 @@
-import { getFlag1 } from '../helpers/index.js';
+import { getFlag1 } from '../common/index.js';
 
 fixture `Fixture`;
