+++ test/functional/fixtures/api/es-next/custom-client-scripts/testcafe-fixtures/test-api.js
@@ -1,4 +1,4 @@
-import { getFlag1, getFlag2 } from '../helpers/index.js';
+import { getFlag1, getFlag2 } from '../common/index.js';
 
 fixture `Fixture`
     .clientScripts('../data/set-flag1.js');
